requirements.txt
discord.py
pytz




google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pillow>=10.0.0
pytesseract>=0.3.10